requirements.txt
blockfrost-python @ git+https://github.com/blockfrost/blockfrost-python@e16afc9323f91e67ae4440dc5652b3caf851b3e7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==1.11.23
djangorestframework==3.5.4
glob2==0.7
idna==3.2
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
